Merge From frontend into liveServer

See merge request s09-bigdata-dist-sub2/S09P22B302!337

+++ ttsServer/src/main/java/com/toy/kafka/dto/playList/MusicDto.java
@@ -10,7 +10,7 @@ import java.io.Serializable;
 public class MusicDto implements Serializable {
     private String type;
     private String path;
-    private int length;
+    private long length;
     private String artist;
     private String title;
     private String image;
